+++ docs/conf.py
@@ -1,6 +1,6 @@
 # -*- coding: utf-8 -*-
 DESCRIPTION = (
-    'Yet another jinja2 cli command for static text generation' +
+    'General purpose static text generator' +
     ''
 )
 # Configuration file for the Sphinx documentation builder.
@@ -25,9 +25,9 @@ project = 'moban'
 copyright = '2017-2019 Onni Software Ltd.'
 author = 'C. W.'
 # The short X.Y version
-version = '0.6.8'
+version = '0.7.0'
 # The full version, including alpha/beta/rc tags
-release = '0.6.8'
+release = '0.7.0'
 
 # -- General configuration ---------------------------------------------------