+++ moban/core/hashstore.py
@@ -1,20 +1,15 @@
-import sys
 import json
 import hashlib
 
 from moban import constants
 from moban.externals import file_system
 
-PY2 = sys.version_info[0] == 2
-
 
 class HashStore:
     IGNORE_CACHE_FILE = False
 
     def __init__(self):
-        self.cache_file = file_system.to_unicode(
-            constants.DEFAULT_MOBAN_CACHE_FILE
-        )
+        self.cache_file = constants.DEFAULT_MOBAN_CACHE_FILE
         if (
             file_system.exists(self.cache_file)
             and self.IGNORE_CACHE_FILE is False
@@ -68,13 +63,10 @@ def get_file_hash(afile):
 
 def get_hash(content):
     md5 = hashlib.md5()
-    if PY2 and content.__class__.__name__ == "unicode":
-        content = content.encode("utf-8")
     md5.update(content)
     return md5.digest().decode("latin1")
 
 
 def _mix(content, file_permissions_copy):
-    if not PY2:
-        file_permissions_copy = file_permissions_copy.encode("utf-8")
+    file_permissions_copy = file_permissions_copy.encode("utf-8")
     return content + file_permissions_copy