+++ moban/core/mobanfile/templates.py
@@ -8,35 +8,33 @@ LOG = logging.getLogger(__name__)
 
 
 def handle_template(template_file, output, template_dirs):
-    LOG.info("handling %s" % template_file)
+    LOG.info(f"handling {template_file}")
 
-    template_file = file_system.to_unicode(template_file)
+    template_file = template_file
     multi_fs = file_system.get_multi_fs(template_dirs)
     if template_file.endswith("**"):
         source_dir = template_file[:-3]
         _, fs = multi_fs.which(source_dir)
         if fs:
-            for a_triple in _listing_directory_files_recusively(
+            yield from _listing_directory_files_recusively(
                 fs, source_dir, output
-            ):
-                yield a_triple
+            )
         else:
             if STORE.look_up_by_output.get(template_file) is None:
                 reporter.report_error_message(
-                    "{0} cannot be found".format(template_file)
+                    f"{template_file} cannot be found"
                 )
     else:
         _, fs = multi_fs.which(template_file)
         if fs is None:
             if STORE.look_up_by_output.get(template_file) is None:
                 reporter.report_error_message(
-                    "{0} cannot be found".format(template_file)
+                    f"{template_file} cannot be found"
                 )
             else:
                 yield _create_a_single_target(template_file, output)
         elif fs.isdir(template_file):
-            for a_triple in _list_dir_files(fs, template_file, output):
-                yield a_triple
+            yield from _list_dir_files(fs, template_file, output)
         else:
             yield _create_a_single_target(template_file, output)
 
@@ -47,25 +45,24 @@ def _list_dir_files(fs, source, dest):
         # hence the following statement looks like cross platform
         #  src_file_under_dir = os.path.join(source, file_name)
         # but actually it breaks windows instead.
-        src_file_under_dir = "%s/%s" % (source, file_name)
+        src_file_under_dir = f"{source}/{file_name}"
         if fs.isfile(src_file_under_dir):
-            dest_file_under_dir = dest + "/" + file_name
+            dest_file_under_dir = f"{dest}/{file_name}"
             template_type = _get_template_type(src_file_under_dir)
             yield (src_file_under_dir, dest_file_under_dir, template_type)
 
 
 def _listing_directory_files_recusively(fs, source, dest):
     for file_name in fs.listdir(source):
-        src_file_under_dir = source + "/" + file_name
-        dest_file_under_dir = dest + "/" + file_name
+        src_file_under_dir = f"{source}/{file_name}"
+        dest_file_under_dir = f"{dest}/{file_name}"
         if fs.isfile(src_file_under_dir):
             template_type = _get_template_type(src_file_under_dir)
             yield (src_file_under_dir, dest_file_under_dir, template_type)
         elif fs.isdir(src_file_under_dir):
-            for a_triple in _listing_directory_files_recusively(
+            yield from _listing_directory_files_recusively(
                 fs, src_file_under_dir, dest_file_under_dir
-            ):
-                yield a_triple
+            )
 
 
 def _create_a_single_target(template_file, output):
@@ -73,9 +70,8 @@ def _create_a_single_target(template_file, output):
     # output.jj2: source.jj2 means 'copy'
     if template_type and output.endswith("." + template_type):
         LOG.info(
-            "template type switched to from {0} to {1}".format(
-                template_type, constants.TEMPLATE_COPY
-            )
+            f"template type switched to from {template_type} to "
+            + constants.TEMPLATE_COPY
         )
         template_type = constants.TEMPLATE_COPY
     return (template_file, output, template_type)