+++ moban/externals/file_system.py
@@ -17,7 +17,6 @@ except ImportError:
     from urlparse import urlparse
 
 
-PY2 = sys.version_info[0] == 2
 LOG = logging.getLogger(__name__)
 
 path_join = fs.path.join
@@ -27,9 +26,9 @@ path_splitext = fs.path.splitext
 def url_join(path, path2):
     result = urlparse(path)
     if result.scheme and path.endswith(result.scheme):
-        return path + to_unicode("!/") + path2
+        return f"{path}!/{path2}"
     else:
-        return path + to_unicode("/") + path2
+        return f"{path}/{path2}"
 
 
 def log_fs_failure(function_in_this_module):
@@ -37,12 +36,15 @@ def log_fs_failure(function_in_this_module):
         try:
             return function_in_this_module(*args, **kwds)
         except fs.errors.CreateFailed:
-            from moban import reporter
+            from moban.externals import reporter
 
             message = "Failed to open %s" % args[0]
             LOG.debug(message)
             reporter.report_error_message(message)
             raise exceptions.FileNotFound(args[0])
+        except fs.opener.errors.UnsupportedProtocol as e:
+            LOG.exception(e)
+            raise exceptions.UnsupportedPyFS2Protocol(e)
 
     return wrapper
 
@@ -50,7 +52,6 @@ def log_fs_failure(function_in_this_module):
 @log_fs_failure
 @contextmanager
 def open_fs(path):
-    path = to_unicode(path)
     if is_zip_alike_url(path):
         zip_file, folder = url_split(path)
         the_fs = fs.open_fs(zip_file)
@@ -65,7 +66,6 @@ def open_fs(path):
 @log_fs_failure
 @contextmanager
 def open_file(path):
-    path = to_unicode(path)
     if is_zip_alike_url(path):
         zip_file, folder = url_split(path)
         the_fs = fs.open_fs(zip_file)
@@ -85,7 +85,6 @@ def open_file(path):
 @log_fs_failure
 @contextmanager
 def open_binary_file(path):
-    path = to_unicode(path)
     if is_zip_alike_url(path):
         zip_file, folder = url_split(path)
         the_fs = fs.open_fs(zip_file)
@@ -120,7 +119,6 @@ read_text = read_unicode
 
 @log_fs_failure
 def write_bytes(filename, bytes_content):
-    filename = to_unicode(filename)
     if "://" in filename:
         zip_file, folder = url_split(filename)
         with fs.open_fs(zip_file, create=True) as the_fs:
@@ -146,10 +144,7 @@ def is_file(path):
         return the_fs.isfile(path)
 
 
-@log_fs_failure
 def exists(path):
-    path = to_unicode(path)
-
     if is_zip_alike_url(path):
         zip_file, folder = url_split(path)
         try:
@@ -159,6 +154,9 @@ def exists(path):
                 return True
         except fs.errors.CreateFailed:
             return False
+        except fs.opener.errors.UnsupportedProtocol as e:
+            LOG.exception(e)
+            raise exceptions.UnsupportedPyFS2Protocol(e)
     dir_name = fs.path.dirname(path)
     the_file_name = fs.path.basename(path)
 
@@ -167,11 +165,13 @@ def exists(path):
             return the_fs.exists(the_file_name)
     except fs.errors.CreateFailed:
         return False
+    except fs.opener.errors.UnsupportedProtocol as e:
+        LOG.exception(e)
+        raise exceptions.UnsupportedPyFS2Protocol(e)
 
 
 @log_fs_failure
 def list_dir(path):
-    path = to_unicode(path)
     folder_or_file, path = _path_split(path)
     with fs.open_fs(folder_or_file) as the_fs:
         for file_name in the_fs.listdir(path):
@@ -180,7 +180,6 @@ def list_dir(path):
 
 @log_fs_failure
 def abspath(path):
-    path = to_unicode(path)
     folder_or_file, path = _path_split(path)
     with fs.open_fs(folder_or_file) as the_fs:
         return the_fs.getsyspath(path)
@@ -188,7 +187,6 @@ def abspath(path):
 
 @log_fs_failure
 def fs_url(path):
-    path = to_unicode(path)
     folder_or_file, path = _path_split(path)
     with fs.open_fs(folder_or_file) as the_fs:
         return the_fs.geturl(path, purpose="fs")
@@ -196,15 +194,12 @@ def fs_url(path):
 
 @log_fs_failure
 def system_path(path):
-    path = to_unicode(path)
     folder_or_file, path = _path_split(path)
     with fs.open_fs(folder_or_file) as the_fs:
         return the_fs.getsyspath(path)
 
 
 def to_unicode(path):
-    if PY2 and path.__class__.__name__ != "unicode":
-        return u"".__class__(path)
     return path
 
 
@@ -250,7 +245,6 @@ def url_split(url):
 
 
 def _path_split(url_or_path):
-    url_or_path = to_unicode(url_or_path)
     if is_zip_alike_url(url_or_path):
         return url_split(url_or_path)
     else: