+++ moban/main.py
@@ -4,7 +4,7 @@
 
     Bring jinja2 to command line
 
-    :copyright: (c) 2016-2019 by Onni Software Ltd.
+    :copyright: (c) 2016-2020 by Onni Software Ltd.
     :license: MIT License, see LICENSE for more details
 
 """
@@ -32,9 +32,6 @@ LOG_LEVEL = [logging.ERROR, logging.WARNING, logging.INFO, logging.DEBUG]
 
 
 def main():
-    """
-    program entry point
-    """
     parser = create_parser()
     options = vars(parser.parse_args())
     handle_verbose(options[constants.LABEL_VERBOSE])
@@ -58,6 +55,7 @@ def main():
             exceptions.DirectoryNotFound,
             exceptions.NoThirdPartyEngine,
             exceptions.MobanfileGrammarException,
+            exceptions.UnsupportedPyFS2Protocol,
         ) as e:
             LOG.exception(e)
             reporter.report_error_message(str(e))
@@ -66,7 +64,10 @@ def main():
         try:
             count = handle_command_line(options)
             moban_exit(options[constants.LABEL_EXIT_CODE], count)
-        except exceptions.NoTemplate as e:
+        except (
+            exceptions.NoTemplate,
+            exceptions.UnsupportedPyFS2Protocol,
+        ) as e:
             reporter.report_error_message(str(e))
             moban_exit(options[constants.LABEL_EXIT_CODE], constants.ERROR)
 
@@ -81,9 +82,6 @@ def moban_exit(exit_code_toggle_flag, exit_code):
 
 
 def create_parser():
-    """
-    construct the program options
-    """
     parser = argparse.ArgumentParser(
         prog=constants.PROGRAM_NAME, description=constants.PROGRAM_DESCRIPTION
     )
@@ -108,13 +106,13 @@ def create_parser():
     )
     advanced.add_argument(
         "-td",
-        "--%s" % constants.LABEL_TMPL_DIRS,
+        f"--{constants.LABEL_TMPL_DIRS}",
         nargs="*",
         help="add more directories for template file lookup",
     )
     advanced.add_argument(
         "-cd",
-        "--%s" % constants.LABEL_CONFIG_DIR,
+        f"--{constants.LABEL_CONFIG_DIR}",
         help="the directory for configuration file lookup",
     )
     advanced.add_argument(
@@ -124,12 +122,12 @@ def create_parser():
         "-g", "--%s" % constants.LABEL_GROUP, help="a subset of targets"
     )
     advanced.add_argument(
-        "--%s" % constants.LABEL_TEMPLATE_TYPE.replace("_", "-"),
+        f"--{constants.LABEL_TEMPLATE_TYPE.replace('_', '-')}",
         help="the template type, default is jinja2",
     )
     advanced.add_argument(
         "-d",
-        "--%s" % constants.LABEL_DEFINE,
+        f"--{constants.LABEL_DEFINE}",
         nargs="+",
         help=(
             "to supply additional or override predefined variables,"
@@ -138,7 +136,7 @@ def create_parser():
     )
     advanced.add_argument(
         "-e",
-        "--%s" % constants.LABEL_EXTENSION,
+        f"--{constants.LABEL_EXTENSION}",
         nargs="+",
         help="to to TEMPLATE_TYPE=EXTENSION_NAME",
     )
@@ -153,7 +151,7 @@ def create_parser():
         "Developer options", "For debugging and development"
     )
     developer.add_argument(
-        "--%s" % constants.LABEL_EXIT_CODE,
+        f"--{constants.LABEL_EXIT_CODE}",
         action="store_true",
         dest=constants.LABEL_EXIT_CODE,
         default=False,
@@ -161,7 +159,7 @@ def create_parser():
     )
     developer.add_argument(
         "-V",
-        "--%s" % constants.LABEL_VERSION,
+        f"--{constants.LABEL_VERSION}",
         action="version",
         version="%(prog)s {v}".format(v=__version__),
     )